ANTLR4 support
(I haven't been using this lately…may want to revert at some point?)

(Also: what the heck is the deal with tabtab and electron-forge? Is any of this still relevant, or was I just playing around and forgot to revert? Probably the latter, but this commit at least gives me insurance. 🙄)