P.A.R.A. CLIs
> Rhymes with Peracles?

The `para` function is tailored to Tiago Forte's P.A.R.A. system: https://praxis.fortelabs.co/the-p-a-r-a-method-a-universal-system-for-organizing-digital-information

`para init` generates the requisite 4 folders, if they don't already exist. I'm tempted to also have it move every other directory in `.` into `4-Archives`, but that's a feature for another day.

`para p` cd's to `1-Projects`.
`para a` cd's to `2-Areas of Responsibility folder`.
`para r` cd's to `3-Resources`.
`para b` cde's to `4-Archives`.

Or you can add one or more paths to have them `mv`'d to the corresponding folder. For example…

```
$ pwd
/Users/c/Documents/3-Resources
$ para p ~/Desktop/foo
```

Has the effect of moving `~/Desktop/foo` to `/Users/c/Documents/1-Projects/foo`.  I'm toying with having it _also_ `cd` to `1-Projects`.

Also up for consideration: setting a PARA_ROOT environment variable so that `para` can be invoked successfully from any directory, rather than the current "root-or-child-or-bust" approach.