New functions: cod & nt
`cod` is `code` without the `e`, which is to say, it's a super trivial way to quickly open the current directory in both VS Code and SourceTree.

`nt` stands for "New Tab". It opens a new Terminal.app tab & automatically `cd`'s to the working directory from which it was run. (I wish Terminal.app automatically did this with Fish shell…but it doesn't. ☹️)