Print the colors listed by `set_color --print-colors` IN THE COLOR SPECIFIED!
Why set_color doesn't offer this option is a mystery to me.

Exercise for another time, perhaps: modify this to manipulate the output of `set_color --print-colors` in real time so that it's more robust. (There's not much likelihood that the colors will change much between now and whenever, but this would be a good way to practice shell manipulations of lists and strings.)