`dora l` now omits listing the 4-Archives folder. (Use `dora ll` instead.)
My Archives are starting to get crowded, polluting the output of the original `dora l`.

A future improvement might be to offer even more customization of the listing commands, so I can list, for example, just what's in 1-Due. Or perhaps separate these out to a separate function/set of functions? Whatever…not worrying about that now.