gpop → git pop (or apply) a named stash
This function makes it trivial to apply a stash repeatedly across branches. Used, for example, for working on roam-toolkit, which currently fails in dev mode without a tweak to manifest.json which can't be committed because it breaks production. (!) I don't care to troubleshoot this issue right now, so I just revert the file whenever I need to have a clean workspace, then `gpop roam` whenever I need the dev version of manifest.json back.